feat: move JavaScript logic for event upload form to a separate file for better organization
