+++ resources/js/app.js
@@ -15,3 +15,4 @@ import "./termsModal.js";
 import "./buttonTerm.js";
 import "./accordion.js";
 import "./modal.js";
+import "./formUploadAdmin.js";
